# candidates-openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "api-companies",
        "description": "Microservices candidates",
        "version": "1.0.0",
        "x-logo": {
            "url": "https://fastapi.tiangolo.com/img/logo-margin/logo-teal.png"
        }
    },
    "paths": {
...
        "/api/candidates/workexperience": {
            "get": {
                "tags": [
                    "workexperience"
                ],
                "summary": "Get Workexperiences",
                "operationId": "get_workexperiences_api_candidates_workexperience_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseDto"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "workexperience"
                ],
                "summary": "Create Workexperiences",
                "operationId": "create_workexperiences_api_candidates_workexperience_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WorkExperience"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseDto"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/candidates/workexperience/{id}": {
            "get": {
                "tags": [
                    "workexperience"
                ],
                "summary": "Get Workexperiences",
                "operationId": "get_workexperiences_api_candidates_workexperience__id__get",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseDto"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "workexperience"
                ],
                "summary": "Put Workexperience",
                "operationId": "put_workexperience_api_candidates_workexperience__id__put",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WorkExperienceUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseDto"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EmploymentType": {
                "type": "string",
                "enum": [
                    "JORNADA COMPLETA",
                    "JORNADA PARCIAL",
                    "CONTRATO TEMPORAL",
                    "PROFESIONAL INDEPENDIENTE",
                    "CONTRATO DE PRÁCTICAS"
                ],
                "title": "EmploymentType"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ResponseDto": {
                "properties": {
                    "message": {
                        "type": "string",
                        "title": "Message"
                    },
                    "data": {
                        "anyOf": [
                            {
                                "items": {},
                                "type": "array"
                            },
                            {
                                "type": "object"
                            },
                            {
                                "type": "string"
                            }
                        ],
                        "title": "Data"
                    },
                    "errors": {
                        "type": "string",
                        "title": "Errors"
                    }
                },
                "type": "object",
                "required": [
                    "message",
                    "data",
                    "errors"
                ],
                "title": "ResponseDto"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
            "WorkExperience": {
                "properties": {
                    "company": {
                        "type": "string",
                        "title": "Company"
                    },
                    "job_title": {
                        "type": "string",
                        "title": "Job Title"
                    },
                    "start_date": {
                        "type": "string",
                        "title": "Start Date"
                    },
                    "end_date": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "End Date"
                    },
                    "functions": {
                        "type": "string",
                        "title": "Functions"
                    },
                    "job_type": {
                        "$ref": "#/components/schemas/EmploymentType"
                    },
                    "is_actual": {
                        "type": "boolean",
                        "title": "Is Actual"
                    },
                    "candidate_id": {
                        "type": "integer",
                        "title": "Candidate Id"
                    }
                },
                "type": "object",
                "required": [
                    "company",
                    "job_title",
                    "start_date",
                    "functions",
                    "job_type",
                    "is_actual",
                    "candidate_id"
                ],
                "title": "WorkExperience"
            },
            "WorkExperienceUpdate": {
                "properties": {
                    "company": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Company"
                    },
                    "job_title": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Job Title"
                    },
                    "start_date": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Start Date"
                    },
                    "end_date": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "End Date"
                    },
                    "functions": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Functions"
                    },
                    "job_type": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/EmploymentType"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "is_actual": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Is Actual"
                    },
                    "candidate_id": {
                        "type": "integer",
                        "title": "Candidate Id"
                    }
                },
                "type": "object",
                "required": [
                    "candidate_id"
                ],
                "title": "WorkExperienceUpdate"
            }
        }
    }
}